amd64: barely boot into kernel code

+++ src/kernel/arch/amd64/registers.h
@@ -3,19 +3,16 @@
 #include <stdint.h>
 
 struct registers {
-	/* those are in the order of pushad/popad - so you can load/save this
-	 * struct in (almost) one instruction */
-	uint32_t edi, esi;
+	uint64_t edi, esi;
 	userptr_t ebp, esp;
-	uint32_t ebx, edx, ecx, eax;
+	uint64_t ebx, edx, ecx, eax;
 
-	// not part of pushad/popad, but still useful
 	userptr_t eip;
 } __attribute__((__packed__));
 
 // saves a return value according to the SysV ABI
 static inline uint64_t regs_savereturn(struct registers *regs, uint64_t value) {
 	regs->eax = value;
-	regs->edx = value >> 32;
+	regs->edx = value >> 32; // TODO check ABI
 	return value;
 }
